requirements.txt
google-genai
Pillow
Flask
python-dotenv
google-cloud-storage
google-cloud-aiplatform
Flask-Cors